+++ packages/react-components/react-message-bar-preview/src/components/MessageBarBody/MessageBarBody.types.ts
@@ -7,7 +7,7 @@ export type MessageBarBodySlots = {
 /**
  * MessageBarBody Props
  */
-export type MessageBarBodyProps = ComponentProps<MessageBarBodySlots> & {};
+export type MessageBarBodyProps = ComponentProps<MessageBarBodySlots>;
 
 /**
  * State used in rendering MessageBarBody